Make i2c and spi remove callbacks return void
The upstream commit ed5c2f5fd10d ("i2c: Make remove callback return
void") and the upstream commit a0386bba7093 ("spi: make remove callback
a void function") change the return value for the i2c_device and
spi_device remove callbacks to be void. So we need to update all of
these drivers accordingly. The reason was because the return value
wasn't really being used which is consistent with this driver as well.

Bug: 253041218
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: I5fafbed4ae7edea8dd316cd05fea09405efe7a4f