LWIS: Add a struct device field in struce lwis_device
The plat_dev in lwis_device might be invalid when the device type
is not platform driver. But there are many common codes for all
device types that access plat_dev. It will cause NPE when the
device type is not a platform driver. Most of the codes to access
plat_dev are to access the struct device in it. So, we can add a
new struct device field in struct lwis_device and assign the
struct device value from struct platform_device when probe. Then
use the new added field in the common code. It should fix the NPE
when the device type is not a platform driver.

Bug: 278156602
Test: GCA, CTS
Change-Id: Ic46b5ea66d85e1299516c8da582e06ae69eedb41
Signed-off-by: Holmes Chou <holmeschou@google.com>
Signed-off-by: Nick Chung <nickchung@google.com>
(cherry picked from commit b5e90bd1462896158a9d13f97dcb831661fedb0f)