panel: added pps buffer write as a function
Writing out the `struct drm_dsc_config` data as a `pps` buffer is common
to a number of panels; as such, this change centralizes that behavior.

The conditional compilation flag around this function relates to the
buffer-writing helper changing location in later versions of the kernel,
as well as being contingent on the CONFIG_DRM_DISPLAY_DP_HELPER flag.

Bug: 262417081
Test: none
Change-Id: I8d6bcf935e70ac180753f23ba853a691b16e77b5
Signed-off-by: Taylor Nelms <tknelms@google.com>