bigwave: Add interface to skip cache operations on map/unmap
Map/unmap performance cache operations on whole buffer which is not
always required. Performance can be improved if these cache operataions
are done only on the portion of buffer which contans relevant data. So,
add a interface to skip cache operations on map/unmap. If cache
operations are required then client can explicitly call partial cache
operations for better performance.

bug: 279537928

Change-Id: I0bb72844f0d20f7485a7344a1d8298e1532573d6
Signed-off-by: Vinay Kalia <vinaykalia@google.com>