panel: allow to update panel idle mode later
Updating the panel idle mode in the composer thread may be blocked by
the display driver when a frame is under transfer, resulting in frame
drop. To avoid this, allow the panel idle mode to be updated at the
pre-commit of the next frame, which ensures that command can be sent
without delay.

Bug: 263362191
Test: build
Change-Id: Ia2345da53919e9b4b5479af64dd8704327fff1a8
Signed-off-by: Midas Chien <midaschieh@google.com>