IoReg: Move out iommu_dma_enable_best_fit_algo() call
When we call iommu_dma_enable_best_fit_algo() in the
lwis_ioreg_device_parse_dt(), it needs a struct device input. We are
using ioreg_dev->base_dev.dev as input now, but it is not ready at this
moment. The best place to call iommu_dma_enable_best_fit_algo() should
be in lwis_ioreg_device_setup().

Bug: 321131062
Test: GCA
Change-Id: I257dfd3350500d21f9480fddebc4dec2af2004e8
Signed-off-by: Holmes Chou <holmeschou@google.com>