google-modules: Only use edgetpu's gcip driver when modules are disabled
The gcip-kernel-driver folders between edgetpu and gxp are almost the same,
but the edgetpu one is newer and has required IOMMU changes, so just build
the edgetpu one instead of the gxp one when modules are disabled. The
edgetpu version appears to be compatible with gxp too, so all is well.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>