Use `dma_fence` callbacks for transaction triggering
Instead of having our own list in `lwis_fence`, we can use `dma_fence`
list by adding one callback for each transaction that's going to be
triggered.

This patch also makes sure that no callbacks are left in the `dma_fence`
list if the transactions were canceled. This avoids complains from the
`dma_fence` release code when a fence is being released with pending
callbacks.

Test: liblyric.lwis_test
Bug: 342031592
Change-Id: I1cdd5e5bb364ddef5626cbe468ec463119aebe9f
Signed-off-by: Luis Useche <useche@google.com>