Remove `lwis_fence_get_status[_locked]` functions
These are now replaced by the corresponding functions in DMA fences.

With this patch we complete the transition to DMA fences in LWIS. From
this point on, user space can send DMA fence file descriptors in the
fence file descriptors and LWIS will handle it appropriately. This
include LWIS fences and DMA fences coming from outside of the Lyric
world.

Test: liblyric.lwis_test
Bug: 342032857
Change-Id: I3b35e537adc7b2b542a24fb5d4e9999c1c7383da
Signed-off-by: Luis Useche <useche@google.com>