edgetpu: Fix potential etdomains race condition
As of the change supporting 16K pages, we can't guarantee that mapping
of SCID to PASID is 1:1. Even with 4K pages, it seems we can't guarantee
that according to the experiment.

That says, before the 16K page support, domains in each index (PASID) of
the `etiommu->etdomains` array always had the same unique SCID and each
client could simply fetch its domain according to its PASID without any
race conditions.

However, that's not true anymore and different domains with the same
SCID can be assigned to the different index (PASID) of the array without
this change. We should guarantee that domains in the array to always
have different context ID (SCID).

Bug: 343896107
Test: With cl/641706538, verified that every repeating ./run_kernel_test
      running in 6 terminals concurrently are passing well over 6 hrs.
Change-Id: I588f3132f336f15e077392f9ecca552d6cf983c7
Signed-off-by: Junghoon Jang <junghoonjang@google.com>