panel: hk3: allow using different settings for sub-120Hz mode
There are display artifacts when using higher PWM frequency at lower
refresh rate. This commit introduces another set of commands when
refresh rate is lower than 120Hz (or NS mode - I'm not sure if NS mode
also requires this new command, but looks like it)

Default values are not changed. This commit only provides the possibility
to do so.