zuma_defconfig: Enable CONFIG_PM_USERSPACE_AUTOSLEEP
Android goes into suspend aggressively when the screen is turned off. This
is different from the behavior of desktop and laptop PCs, and necessitates
the kernel knowing that rapid suspend entry/exit is going on.

Enable CONFIG_PM_USERSPACE_AUTOSLEEP for that purpose, which alters
behavior in WireGuard and the CRNG to make them play well with Android.

This indication was previously covered by CONFIG_ANDROID in older kernels.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>